Generalised cookie function to browsers/utils.go
